# profile: comprehensive Lambda startup performance analysis