## Summary
Add detailed profiling to identify Lambda startup bottlenecks causing 9.5s cold start and `expired_trigger_id` errors.

## Problem
Current Lambda performance:
- **Duration**: 9,497ms (9.5 seconds) 
- **Init Duration**: 2,052ms (2 seconds)
- **Error**: `expired_trigger_id` - exceeds Slack's 3-second timeout
- **User Impact**: Modal never appears when clicking "emoji-smith" action

## Profiling Added
**Lambda Handler** (`src/lambda_handler.py`):
- App initialization timing
- Import vs execution timing breakdown

**App Creation** (`src/emojismith/app.py`):
- Step-by-step webhook handler creation:
  - Environment setup
  - Slack client creation  
  - OpenAI client creation
  - Image processor initialization
  - Domain services creation
  - Job queue setup (SQS/aioboto3 import timing)
  - Security service creation
- FastAPI app creation timing

## Expected Data
This will help identify the slowest components:
- Third-party library imports (slack-sdk, openai, PIL)
- Client initialization overhead
- Dependency injection bottlenecks
- Potential optimization targets

## Next Steps
1. Deploy and test to capture CloudWatch profiling logs
2. Analyze timing data to identify optimization targets
3. Implement targeted optimizations based on data
4. Consider immediate ACK pattern if startup can't be optimized enough

🤖 Generated with [Claude Code](https://claude.ai/code)